+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 * Added property `Polygon.width`.
 * [mildly breaking] Changed the output of `Polygon.clip_out_of_image()` from `MultiPolygon` to `list` of `Polygon`.
   This breaks for anybody who has already used `Polygon.clip_out_of_image()`. 
+* Fixed `Affine` heatmap augmentation producing arrays with values outside the range `[0.0, 1.0]` when `order` was set to `3`.
 
 
 # 0.2.8